+++ packages/web-components/vite.jsdelivr-config.ts
@@ -3,6 +3,10 @@ import dts from "vite-plugin-dts";
 
 import { dependencies } from "./vite.base-config";
 
+import clientPkg from "../storefront-client/package.json" assert { type: "json" };
+import cartPkg from "../storefront-cart/package.json" assert { type: "json" };
+import searchPkg from "../predictive-search/package.json" assert { type: "json" };
+
 const JS_DELIVR_URL_BASE = "/npm/";
 
 export default defineConfig({
@@ -15,11 +19,24 @@ export default defineConfig({
     rollupOptions: {
       output: {
         paths: (id) => {
-          if (dependencies[id] === "workspace:*") return id;
+          if (dependencies[id] === "workspace:*") {
+            const version = dependencies[id]?.replace(
+              "workspace:*",
+              clientPkg.name === id
+                ? clientPkg.version
+                : cartPkg.name === id
+                  ? cartPkg.version
+                  : searchPkg.name === id
+                    ? searchPkg.version
+                    : ""
+            );
+            if (!version) return id;
+            return `${JS_DELIVR_URL_BASE}${id}@${version}`;
+          }
+
           const version = dependencies[id]?.replace("^", "@");
           if (!version) return id;
-          const url = `${JS_DELIVR_URL_BASE}${id}${version}/+esm`;
-          return url;
+          return `${JS_DELIVR_URL_BASE}${id}${version}/+esm`;
         },
       },
     },